# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ISP Billing API
  version: 1.0.0
servers:
  - url: https://your-host/api
paths:
...
  /audit-logs:
    get:
      summary: List tenant audit events with filters, pagination, and redacted payloads
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: actor
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 250 }
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
  /audit-logs/actions:
    get:
      summary: List tenant audit action names with counts for filters
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
  /audit-logs/summary:
    get:
      summary: Summarize matching tenant audit events by action and actor
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
...
  /api-keys:
    get:
      summary: List tenant API keys without exposing secret material
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
    post:
      summary: Create a scoped tenant API key and return the plaintext key once
      security:
        - bearerAuth: []
      responses:
        '201': { description: Created }
        '400': { description: Invalid scopes or expiry }
        '403': { description: Owner or admin role required }
  /api-keys/scopes:
    get:
      summary: List supported tenant API key scopes
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
  /api-keys/{id}:
    patch:
      summary: Update API key label, description, scopes, or expiry
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: API key not found }
        '409': { description: Revoked API key cannot be updated }
    delete:
      summary: Revoke a tenant API key
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: API key not found }
...
  /integration/v1/customers:
    get:
      summary: List tenant customers through a scoped integration API key
      security:
        - apiKeyAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        '200': { description: OK }
        '403': { description: customers:read scope required }
  /integration/v1/customers/{id}:
    get:
      summary: Get one tenant customer through a scoped integration API key
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: customers:read scope required }
        '404': { description: Customer not found }
...
  /invoices:
    get:
      summary: List tenant invoices with finance status and balances
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
  /invoices/{id}:
    get:
      summary: Get a finance audit view for one invoice including allocations, source credits, and ledger entries
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
...
  /support/tickets:
    get:
      summary: List support tickets with linked customers, assets, and work orders
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
    post:
      summary: Create a support ticket with SLA timestamps and optional linkage
      security:
        - bearerAuth: []
      responses:
        '201': { description: Created }
        '403': { description: Forbidden }
...
  /platform-api/gateway-events/tenants/{tenantId}/customers/search:
    get:
      summary: Search customer candidates inside a tenant for manual orphan resolution
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tenantId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
  /customers/{id}/portal-access:
    get:
      summary: Get admin-safe customer portal access status without exposing PIN hash material
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Customer not found }
    put:
      summary: Enable or disable customer portal access, set a portal PIN, or clear the PIN
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isEnabled: { type: boolean }
                pin: { type: string, minLength: 4, maxLength: 8 }
                clearPin: { type: boolean }
                reason: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Invalid request or PIN format }
        '403': { description: Owner or admin role required }
        '404': { description: Customer not found }
  /customers/{id}/communication-preferences:
    get:
      summary: Get customer messaging consent, suppression, and quiet-hours preferences
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Customer not found }
    put:
      summary: Update customer messaging consent, suppression, and quiet-hours preferences
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preferredLanguage: { type: string }
                transactionalSmsEnabled: { type: boolean }
                billingSmsEnabled: { type: boolean }
                serviceAlertsSmsEnabled: { type: boolean }
                marketingSmsEnabled: { type: boolean }
                doNotContactUntil: { type: string, format: date-time, nullable: true }
                quietHours:
                  type: object
                  properties:
                    enabled: { type: boolean }
                    start: { type: string, example: '21:00' }
                    end: { type: string, example: '07:00' }
                    timezone: { type: string, example: Africa/Nairobi }
      responses:
        '200': { description: OK }
        '400': { description: Invalid preference payload }
        '403': { description: Owner or admin role required }
        '404': { description: Customer not found }
    patch:
      summary: Partially update customer messaging consent, suppression, and quiet-hours preferences
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Invalid preference payload }
        '403': { description: Owner or admin role required }
        '404': { description: Customer not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY